Added ID_NETMESSAGE to the RemoteDialog resources. Gave the group box an ID so it could be hidden. Moved various controls around in that dialog.
All other changes seem to be due to a newer version of the RC.exe compiler.
Most are removal of spaces at ends of lines, but one in particular (changing
the CRichEdit ctrl from "RichEdit20A" to "RichEdit20W" seems like it may be
a problem. I've left it out for now.